backend/requirements.txt
Django>=3.2,<4.0
psycopg2-binary
djangorestframework
gunicorn
django-cors-headers
djangorestframework-simplejwt
django-mptt
django-ckeditor
Pillow
celery
redis
openai